fix mapper # detection
per header https://wiki.nesdev.com/w/index.php/INES , calculation should shift header 7 nibble and then add header 6 nibble. 
An example is the Zelda rom, it uses mapper 1, but the code as written detects it as mapper 16 because it is adding 0 to 1 and shifting the result by 4, instead of shifting 0 by 4 and adding it to 1.